Refactor: modal, carousel, breadcrumb 컴포넌트들을 components/common 폴더로 이동
