+++ client/src/components/common/carousel/carousel-item.tsx
