+++ client/src/components/common/carousel/carousel.tsx
