+++ client/src/components/common/carousel/index.ts
