+++ client/src/components/common/index.tsx
@@ -8,3 +8,6 @@ export {
 export { Radio, RadioGroup } from './radio';
 export { Checkbox } from './checkbox';
 export { Section } from './section';
+export { Breadcrumb } from './breadcrumb';
+export { Carousel, CarouselItem } from './carousel';
+export { Modal, ConfimationModal } from './modal';
